requirements.txt
psycopg2-binary
retrying
ddtrace==0.12.1
